Refactor code and optimize for video call
