+++ src/features/call/components/index.tsx
@@ -1,13 +1,13 @@
 'use client';
 
 import { useEffect } from 'react';
-import ReceiveVideoCall from '../receive-call';
-import VideoCallDragWrapper from './video-call-drag-wrapper';
 import socket from '@/lib/socket-io';
 import { sendEvent } from '../utils/custom-event.util';
 import { SOCKET_CONFIG } from '@/configs/socket';
 import { useVideoCallStore } from '../store/video-call.store';
 import { CommonComponent } from './common/common';
+import VideoCall from '../video-call';
+import ReceiveVideoCall from '../receive-call';
 
 const CallVideoModalContainer = () => {
   const { removeRequestCall } = useVideoCallStore();
@@ -27,7 +27,7 @@ const CallVideoModalContainer = () => {
   }, []);
   return (
     <>
-      <VideoCallDragWrapper />
+      <VideoCall />
       <ReceiveVideoCall />
       <CommonComponent />
     </>
