+++ src/features/call/video-call/components/actions/action-toggle-camera.tsx
@@ -1,0 +1,32 @@
+import { Button } from '@/components/actions';
+import { useMyVideoCallStore } from '@/features/call/store/me.store';
+import { Video, VideoOff } from 'lucide-react';
+import React, { memo } from 'react';
+
+interface ActionToggleCameraProps {
+  handleChangeCameraOrMic: (params: {
+    video?: boolean;
+    audio?: boolean;
+  }) => void;
+}
+const ActionToggleCamera = ({ handleChangeCameraOrMic }: ActionToggleCameraProps) => {
+  const { isTurnOnCamera, setTurnOnCamera } = useMyVideoCallStore();
+  const onToggleCamera = () => {
+    setTurnOnCamera(!isTurnOnCamera);
+    handleChangeCameraOrMic({
+      video: !isTurnOnCamera,
+    });
+  };
+  return (
+    <Button.Icon
+      variant="default"
+      size="xs"
+      color={isTurnOnCamera ? 'primary' : 'default'}
+      onClick={onToggleCamera}
+    >
+      {isTurnOnCamera ? <Video /> : <VideoOff />}
+    </Button.Icon>
+  );
+};
+
+export default memo(ActionToggleCamera);
