+++ src/features/call/video-call/components/actions/dropdown-actions.tsx
@@ -1,0 +1,44 @@
+import { Button } from '@/components/actions';
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuTrigger,
+} from '@/components/data-display';
+import { CALL_TYPE } from '@/features/call/constant/call-type';
+import { useParticipantVideoCallStore } from '@/features/call/store/participant.store';
+import { useVideoCallStore } from '@/features/call/store/video-call.store';
+import { MoreVertical, UserPlus2 } from 'lucide-react';
+import ActionToggleCaption from './action-toggle-caption';
+import ActionDoodle from './action-doodle';
+import ActionToggleLayout from './action-toggle-layout';
+
+export default function DropdownActions() {
+  const { room, isFullScreen, setModalAddUser } = useVideoCallStore();
+
+  return (
+    <DropdownMenu>
+      <DropdownMenuTrigger asChild>
+        <Button.Icon
+          variant="default"
+          size="xs"
+          color="default"
+          className={`${!isFullScreen ? 'hidden' : ''}`}
+        >
+          <MoreVertical />
+        </Button.Icon>
+      </DropdownMenuTrigger>
+      <DropdownMenuContent>
+        <ActionToggleLayout />
+        <ActionDoodle />
+        {room.type === CALL_TYPE.GROUP && (
+          <DropdownMenuItem onClick={() => setModalAddUser(true)}>
+            <UserPlus2 />
+            <span className="ml-2">Add member</span>
+          </DropdownMenuItem>
+        )}
+        <ActionToggleCaption />
+      </DropdownMenuContent>
+    </DropdownMenu>
+  );
+}
