+++ src/features/call/video-call/hooks/use-have-share-screen.ts
@@ -1,0 +1,13 @@
+import { useMemo } from 'react';
+import { useParticipantVideoCallStore } from '../../store/participant.store';
+
+export default function useHaveShareScreen() {
+  const participants = useParticipantVideoCallStore(
+    (state) => state.participants,
+  );
+  const haveShareScreen = useMemo(() => {
+    return participants.some((p) => p.isShareScreen);
+  }, [participants]);
+
+  return haveShareScreen;
+}
